fix(compiler): extract block result validation
