+++ packages/compiler/src/instructionCompilers/else.ts
@@ -2,6 +2,7 @@ import { WASMInstruction } from '@8f4e/compiler-wasm-utils';
 import { BLOCK_TYPE } from '@8f4e/compiler-spec';
 import { ErrorCode } from '@8f4e/compiler-spec';
 
+import consumeExpectedBlockResult from './utils/consumeExpectedBlockResult';
 import { saveByteCode } from './utils/saveByteCode';
 
 import { getError } from '../compilerError';
@@ -19,21 +20,7 @@ const _else: InstructionCompiler = (line, context) => {
 		throw getError(ErrorCode.MISSING_BLOCK_START_INSTRUCTION, line, context);
 	}
 
-	if (block.hasExpectedResult) {
-		const operand = context.stack.pop();
-
-		if (!operand) {
-			throw getError(ErrorCode.INSUFFICIENT_OPERANDS, line, context);
-		}
-
-		if (block.expectedResultIsInteger && !operand.isInteger) {
-			throw getError(ErrorCode.ONLY_INTEGERS, line, context);
-		}
-
-		if (!block.expectedResultIsInteger && operand.isInteger) {
-			throw getError(ErrorCode.ONLY_FLOATS, line, context);
-		}
-	}
+	consumeExpectedBlockResult(block, line, context, { validateFloatResult: true });
 
 	context.blockStack.push(block);
