+++ packages/compiler/src/instructionCompilers/loopEnd.ts
@@ -2,6 +2,7 @@ import { WASMInstruction, br } from '@8f4e/compiler-wasm-utils';
 import { BLOCK_TYPE } from '@8f4e/compiler-spec';
 import { ErrorCode } from '@8f4e/compiler-spec';
 
+import consumeExpectedBlockResult from './utils/consumeExpectedBlockResult';
 import { saveByteCode } from './utils/saveByteCode';
 
 import { getError } from '../compilerError';
@@ -19,19 +20,7 @@ const loopEnd: InstructionCompiler = (line, context) => {
 		throw getError(ErrorCode.MISSING_BLOCK_START_INSTRUCTION, line, context);
 	}
 
-	if (block.hasExpectedResult) {
-		const operand = context.stack.pop();
-
-		if (!operand) {
-			throw getError(ErrorCode.INSUFFICIENT_OPERANDS, line, context);
-		}
-
-		if (block.expectedResultIsInteger && !operand.isInteger) {
-			throw getError(ErrorCode.ONLY_INTEGERS, line, context);
-		}
-
-		context.stack.push(operand);
-	}
+	consumeExpectedBlockResult(block, line, context, { restore: true });
 
 	return saveByteCode(context, [...br(0), WASMInstruction.END, WASMInstruction.END]);
 };
